Rename resource files and move to resources folder
