+++ xlive/resources/resource.h
