+++ xlive/resources/version_resource.h
